% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_isas.R
\name{process_isas}
\alias{process_isas}
\title{Process data for ISAS analysis}
\usage{
process_isas(data)
}
\arguments{
\item{trial_data}{dataframe of other variables}
}
\description{
Iowa Satisfaction with Anesthesia Scale
}
